refactor(ispx): convert from nested module to sub-package
Convert `ispx` from a nested Go module (`github.com/goplus/spx/ispx`) to
a sub-package of the main module (`github.com/goplus/spx/v2/pkg/ispx`).
This simplifies version management and keeps the codebase in sync.

Structure changes:
- Move library code from `ispx/` to `pkg/ispx/`
- Move entry points from `ispx/cmd/ispx/` to `cmd/ispx/`
- Move entry points from `ispx/cmd/ispxpc/` to `cmd/ispxnative/`
- Delete `ispx/go.mod` and `ispx/go.sum` (no longer a separate module)

Import path updates:
- `github.com/goplus/spx/ispx` to `github.com/goplus/spx/v2/pkg/ispx`
- `github.com/goplus/spx/ispx/internal/memfs` to
  `github.com/goplus/spx/v2/pkg/ispx/internal/memfs`

Build script improvements:
- Split `ispx/build.sh` into `cmd/ispx/build.sh` (Wasm) and
  `cmd/ispxnative/build.sh` (native shared library)
- Simplify scripts to only build, letting callers handle installation
- Use `go env GOOS` instead of `uname -s` for reliable OS detection
- Update `cmd/gox/install.sh` to call new build scripts and copy
  artifacts

Dependency changes in root `go.mod`:
- Add `github.com/goplus/ixgo`, `github.com/goplus/mod`,
  `github.com/goplus/reflectx`, `github.com/goplus/xgo`
- Add `tool github.com/goplus/ixgo/cmd/qexp` directive
