+++ cmd/ispx/build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+GOOS=js GOARCH=wasm go build -trimpath -ldflags "-s -w -checklinkname=0" -o ispx.wasm
