+++ cmd/ispxnative/main.go
@@ -8,7 +8,7 @@ import (
 	"path/filepath"
 	_ "unsafe"
 
-	"github.com/goplus/spx/ispx"
+	"github.com/goplus/spx/v2/pkg/ispx"
 )
 
 func main() {
